requirements.txt
langgraph
langchain
langchain-groq
tavily-python
beautifulsoup4
requests
streamlit
python-dotenv
